python/formatting/pyproject.toml
[tool.poetry]
name = "analyser"
version = "0.1.0"
description = ""
authors = ["Dan <danthompson41@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "1.19.2"
pandas = "^1.2.3"
matplotlib = "^3.3.4"
tensorflow = "^2.4.1"
jupyter = "^1.0.0"
black = "^20.8b1"
sklearn = "^0.0"
transformers = "^4.4.2"
tensorflow-hub = "^0.11.0"
tqdm = "^4.59.0"
bert-for-tf2 = "^0.14.9"
pip = "^21.0.1"
tensorflow-text = "^2.4.3"
tf-models-official = "^2.4.0"
pydot = "^1.4.2"
bs4 = "^0.0.1"
beautifulsoup4 = "^4.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
